PORTAL-930: Address strict null checks in most non-React component files (#5896)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-462
Subtask: https://lwd.atlassian.net/browse/PORTAL-930

## 🛠 Changes

- Address errors surfaced by TypeScript's [`strictNullChecks` setting](https://www.typescriptlang.org/tsconfig#strictNullChecks)

## ℹ️ Context for reviewers

View TypeScript's description of [`strictNullChecks` setting](https://www.typescriptlang.org/tsconfig#strictNullChecks) for context on why we need to make these changes. The TLDR is:

> When `strictNullChecks` is `true`, `null` and `undefined` have their own distinct types and you’ll get a type error if you try to use them where a concrete value is expected.

There's like 300 more errors that need addressed, but those mostly fall into a category of "our data model says this field might be null so we need to handle that scenario." — I didn't want to address those in this PR since this is already kinda hard to review. As such, this PR doesn't actually enable `strictNullChecks` yet, but just includes some of the fixes for errors that are surfaced when it is.