PFMLPB-2109: Secure actions for the SaviLinx needed to be updated (#5957)
🎫 Ticket

https://lwd.atlassian.net/browse/PFMLPB-1641
Subtask: https://lwd.atlassian.net/browse/PFMLPB-2304
🛠 Changes

The buttons for Complete, Suppress and Remove are not disabled as planned. Instead, the buttons are enabled, but if the security group does not have access it will pop up a message in the right corner. Each message is different depending on if the button. Adjusted the switching of user roles a bit noticed it was not working all the time.
ℹ️ Context for reviewers

This can only be tested in UAT at the moment, but I used this script for a test plan CPS-3055 already. It passed for the SaviLinx group who do not have access to those buttons (complete, suppress, and remove).

* update the buttons are not disabled, but enable with alert message

* add dependencies and fix issue with  switch roles & depts