PFMLPB-2340: Investigate Preliminary Designation time trigger IDT/ Perf and create a work around (#6028)
🎫 Ticket
https://lwd.atlassian.net/browse/PFMLPB-2340

🛠 Changes
Adjusting where the Preliminary Designation is triggered to prevent any failures during the cypress runs. Triggering it sooner in the process to prevent it from missing during the test later. Added a onTab() to return to the absence hub for these tests to go into the adjudication part immediately after.

ℹ️ Context for reviewers
Tested manually each test to see if each test would pass will run an E2E test suite for this PR to check again.