PORTAL-25: Treat File.name as read-only (#6056)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-25

## 🛠 Changes

- Fix bug where setting `file.name` in some browsers causes a JS error.
- Stop sending a different description for compressed images

## ℹ️ Context for reviewers

`File.prototype.name` is a read-only property.

This Portal functionality was originally intended to help the Contact Center know when a file has been compressed, so they know if the original image was at fault or the Portal compression. This was because when we launched the feature, there were worries that compression might produce poor quality images. That hasn't turned out to the case, so it should be safe to remove this. We can still look in New Relic to identify if an an application had any of its files compressed, if necessary.