PFMLPB-2439: Investigate recent failures in medical_cont_approval_amend_pymt_reduction_90K.ts (DT2) (#6141)
This PR updates a section of the reduction when applying reductions to refocus on the element in order to assert being focus. From testing - there were consistent failures when trying to assert the start date input after typing the date to be focused (see here).

Unfortunately this error was not produceable locally so I ran this test only in CI to replicate. I tried a few different things but ultimately adding the waitForAjaxComplete() and also to focus again once selected - I was able to mitigate the consistent failures.

Co-authored-by: Joseph Garcia <joe@lastcallmedia.com>