create a new test to monitor the Suppress Correspondence (#6160)
https://lwd.atlassian.net/browse/PFMLPB-2438

Created a new test in the unstable to monitor for a few days to see if this test will cause any issues. Found the issue was not showing in my local, but when using the headless browser did the error appear not to click the Notification button in the Options drop-down. Adjust the way we are removing the Suppress Correspondence and checking for the Suppress Correspondence to be disabled in the documents tab.