PFMLPB-2318: Write an E2E test that will select withholding preference via portal (#6180)
🎫 Ticket
https://lwd.atlassian.net/browse/PFMLPB-2316
Subtask: https://lwd.atlassian.net/browse/PFMLPB-2318

🛠 Changes
Updates portal_medical_continuous_approval+Amend_reductionPayments90k.ts so that:

During portal claim submission, the employee elects to have taxes withheld for benefit payments
After approval, split payments are generated for the employee, SIT and FIT.
Uses future leave for claim scenario. (Leave must start in 2022 for generating split payments)
Change in leave type to allow cert doc to be uploaded. This is a requirement to allow the leave admin to review and make amendments to this claim.

ℹ️ Context for reviewers
There were a few smaller adjustments made for this test that are called out in the comments. Most of these changes can be reverted after 1/14/22 to allow for retroactive start dates.