END: Implement a timeout for fineos-poll-deployments (#6239)
Adds a timeout to the fineos-poll-deployments workflow. This workflow should never take more than 2 minutes to complete, even if we trigger a Cypress run during it. There have been a few times recently where snap install was taking ~10 minutes, which means something got seriously hung. We want to prevent that from sticking the job and eating up CI time, so we'll set the timeout to 2 minutes.