PFMLPB-1958: Appeal Acknowledgment appears in Employer Portal (#6172)
This PR adds the ability to view appeal acknowledgment notices in the Employer Portal. To do this, a filter is added to the function to get documents from the FINEOS API. Also, the function to download notices on the employer portal has been updated to download from the appropriate subcase, when the notice is an appeal acknowledgment.