API-2140: Fix DUA Employee Demographics import (#6225)
The `--mode` flag was configured to support multiple values (`nargs="+",`), but
the code was expecting a single mode to be selected. This means the script
wouldn't run anything, as it didn't think any mode was selected. Also rename
this to `--import-mode` for clarity.

If `--start-date` is not provided the docs say we default to processing all
records, but the code was defaulting the value to "now", effectively meaning we
would process _no_ records, unless that run also happened to do an import.
Update the flag default behavior to generally do what we'd want, "now" if
running the import step or "all records" if not running the import step.

Support running just the `SET_ORG_UNIT_ID` step, previously if using mode `FILE`
the code required a `--file` argument, even if not running the step to process a
file.

Add index on `employee.fineos_customer_number` as we are increasingly doing
lookups based on it.

Grab only the most recent `DuaEmployeeDemographics` record for the
Employee+Employer pairs for processing the org unit data. Additional tests for
`SET_ORG_UNIT_ID` step, with associated logic tweaks.

Ticket: https://lwd.atlassian.net/browse/API-2140