INFRA-842: Fix regex for pfml-api services in ECS monitor (#6316)
https://lwd.atlassian.net/browse/INFRA-842

🛠 Changes

    Take numbers into account when ignoring pfml-api-* services.

ℹ️ Context for reviewers

This lambda notifies us whenever an ECS task fails to start. We ignore certain ECS tasks if they run on a frequent schedule or run automatically as an ECS service.

However, right now we're being spammed with notifications due to the pfml-api-trn2 service. Our regex needs to take numbers into account so it knows to ignore this service.