PORTAL-1096: Align on a single variation for responsive tables (#6308)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-1096

## 🛠 Changes

- Removes `responsiveIncludeHeader` and `borderlessMobile` props from `Table` component
- Refactor the custom CSS used for themeing the table.

## ℹ️ Context for reviewers

As part of the new Payment Status page, in #6212 a new stacked table variation was introduced, adding a second visual styling for tables on mobile. The primary differences with this second variation was that it:

- Shows the column label in the first cell
  - Tables can still opt-into hiding this heading after this change, by excluding the `data-label` prop on the `td` element.
- Removes the borders between cells

Rather than have two styles, which complicates the React component's props API, this PR is proposing aligning on one style and sticking to it. The style adopted in this PR is the one introduced in the latest Payment Status work.