API-1984 Use the correct maximum benefit amount in the payments audit (#6394)
## 🎫 Ticket

https://lwd.atlassian.net/browse/api-1984

## 🛠 Changes

Instead of setting the maximum amount for a given week based on the start date of the week, this will set the maximum amount for a given week based on the start date for the claim/absence case. 

## ℹ️ Context for reviewers

Note: the new maximum applies to leave across employers. If there are overlapping claims, the maximum amount is based on the most recent claim start date (which will have the higher maximum amount).