API-2154  Use the IAWW task role for the IAWW task (#6492)
## 🎫 Ticket

https://lwd.atlassian.net/browse/api-2154

## 🛠 Changes

Update the task to use the correct role

## ℹ️ Context for reviewers

After setting up the automated triggers from S3 file creation, the lambda errored out due to a permissions issue. This changes the task to use the specific IAWW role (instead of reusing an existing pub payments role), which matches the trigger and should resolve the permissions issue.