INFA-827: Enable MFA in the infra-test environment (#6441)
* Enable MFA in the infra-test environment

For testing purposes, we want to enable MFA in the infra-test environment first. Some changes to the Cognito pool cannot be reverted, so we need to take extra precautions before merging changes for all environments.

* Fixing IAM roles and policies

The new cognito sms role was not implemented correctly. It needed a policy attachment, assume_role_policy, and an external_id.

Co-authored-by: Matt McCarty <steven.m.mccarty@mass.gov>