PORTAL-1337: Mark employer_dba as Optional/nullable in response models/schemas (#6545)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-1337

## 🛠 Changes

- Updated Pydantic response models to reflect `employer_dba` as `Optional`
- Updated OpenAPI response schemas to reflect `employer_dba` as `nullable`
- Consolidated `EmployerResponse` and `EmployerAddFeinResponse` into a single `EmployerResponse` model since they represent the same shape and intent.

## ℹ️ Context for reviewers

In #5174, the `employer.employer_dba` column was updated to store missing values as `None`/`NULL` rather than an empty string (`""`) in the DB.

One piece in this was missed, which was to update the response models to accurately reflect that this field is now “optional” rather than always a string. For users who are associated with an employer with a `None` value for this field, they will now see an error and be prevented from using the Portal because the data isn't conforming to the Pydantic response model. This PR fixes that.