PORTAL-1390: Support FINEOS multipart document uploads (#6642)
https://lwd.atlassian.net/browse/PORTAL-1390
🛠 Changes

    Add the ENABLE_DOCUMENT_MULTIPART_UPLOAD API flag to switch between base64 and multipart uploads.
        This flag defaults to false and is enabled on TEST, PERF, and CPS-PREVIEW, which are the only environments that currently support multipart uploading in FINEOS.
    Add a new fineos_client.upload_document_multipart method that uploads a document using the FINEOS multipart upload endpoint.

ℹ️ Context for reviewers

When document upload was initially implemented, the multipart upload endpoint was unusable because it corrupted uploads and prevented us from downloading them back again through the FINEOS APIs.

Base64 encodings can bloat file sizes by up to 33%, which reduces the effective maximum file size from 6MB to 4.5MB.

Additionally, in order to support 1099 work, FINEOS needed to enable the Binary Upload flag on their API service. This caused file sizes for claimant uploads to become more bloated, reducing the effective file size from 4.5MB to 3.7MB.

This caused a lot of errors, so FINEOS disabled the Binary Upload flag again in production. However, it will need to be turned on again for the 1099 load on 1/15, and again for periodic incremental loads starting on 2/1.

To raise the effective file size limits back up again, and to support both the 1099 tasks and claimant uploads being available at the same time, we are testing out switching to the multipart upload FINEOS API endpoint.