PFMLPB-2689: Instantiate Azure AD on server start instead of per request (#6488)
Right now Azure AD attempts to instantiate itself if it is not instantiated per request. When the Azure AD related environment variables are not configured, each request will generate a logger.info() entry.

To decrease the noise, it would be better if Azure AD was instantiated when the server starts. The public keys can be switched out over time, so we would still like to refresh the keys every 24 hours.