PFMLPB-2741: Update the Change request approved test with notices/notification checks (#6753)
moved test api_medical_continuous_approval_changeNotice.ts from stable/feature to unstable folder. A few changes were made to the test with the January release coming when a claimant submits a claim through the Portal. The claimant and employer will receive a notification and notice about the new notice called Change Request Approved (PDF) the email template has been updated as well.

Note the email template and FINEOS changes are only in DT2/Test and DT3/cps-preview so far. This is part of the January release