PFMLPB-2494: Convert user from LA to claimant endpoint (#6597)
A self service page will be added to the portal at https://paidleave.mass.gov/user/convert-to-claimant where a user can convert their account from an employer/leave admin to a claimant account if they are not already verified. This adds an endpoint at DELETE /roles that can be used to convert the account from LA to claimant.