API-2265: Investigate how output file is written to S3 - DOR cannot fetch it from the bucket (#6796)
Set off multipart writing for small files. Added a parameter to the write file utilities function to pass a transport parameter to turnoff multipart. Improved test case to detect the incorrect etag value DOR encountered.

It looks like the file was uploaded with multi-part upload, but was small enough to only need one part (hence the -1 at the end of the AWS ETag). When DOR tries to fetch the file, the computed MD5 they generate does not match the ETag in AWS.