API-1929: remove old Application {start/updated}_time columns (#6789)
https://lwd.atlassian.net/browse/API-1929

## Changes
Columns already deprecated but had to stay for a while. Just deleting them now. Looks like there were a couple spots that were still writing to the old values, but nothing reading.

NOTE about the db migration - The downgrade re-adds the columns but won't restore the data. We should already have the desired data in the new created/updated_at columns, and nothing will read from the old columns so it doesn't seem worth it to have the downgrade restore the columns as well.