PORTAL-1504: Fix Employer.has_verification_data not accounting for previously exempt employers (#6837)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-1504

## 🛠 Changes

- Update `Employer.has_verification_data` to use the same constraints as the `employer_get_most_recent_withholding_dates` queries. 

## ℹ️ Context for reviewers

In the Portal, we first check `has_verification_data` in order to determine whether to route an employer to their verification page. On the verification page, we then retrieve the data that the employer should verify.  The logic we use for both is dependent on being consistent, so we should just use the same logic.