PFMLPB-2742: Update the Change Request Denial test with notices/notification checks (#6775)
Updated the api_caring_continuous_approval+denial_changeRequestStatus.ts to include check for notices and notification in the Claimant and Leave Admin Portal. Adjusted the test a bit to have the Changed Request Denial show up. Also adjusted the test for each environment if the env has the FINEOS January release and email template updated will the HAS_FINEOS_JANUARY_RELEASEbe turned to true. The only envs at the moment would be Test, cps-preview, and PERF. If the envs are set up the old way this test will run the old cypress test. This way we are covering this test in all environments somehow.

Adjusted the checkNoticeForLeaveAdmin to remove the switch/case statement.