EDM-405: Certain Doc Types Not Downloading (#6866)
This ticket adds two new notices to the list of documents which are attached to a subcase, and thus cannot be downloaded directly from the absence case. Instead, the specific sub case id must be retrieved before it can be downloaded. Those notices are Max Benefit Amount Change Notice and Benefit Amount Change Notice.

Co-authored-by: James Bursa <james@navapbc.com>