API-2196 Update the evaluate new eligibility report for 2022 changes (#6755)
## 🎫 Ticket

https://lwd.atlassian.net/browse/api-2196

## 🛠 Changes

- Updates the evaluate new eligibility report to run off the main eligibility logic
- Update output files to capture relevant data
- Removes old files we no longer need.

## ℹ️ Context 

We updated the financial eligibility logic to use the leave start date for determining state metrics (e.g. if someone applies in Dec 2021 for leave starting in Jan 2022, their max weekly benefit should be $1084.31, not $850). This change did not make it into PROD until 12/16, so we need to check all FE calls between 11/1 and 12/16 to see how many results would have been different with the new logic. 

This new eligibility report was created after we made similar changes to financial eligibility earlier this year, but needs some updates to work for this specific case.