EDM-404: only show pdf docs (#6867)
Currently, there are three document types that, when first generated in FINEOS, are Word documents. This PR makes sure that only PDF documents are returned by the PFML API to the portal, by filtering using the fileExtension parameter. A test is updated to include this functionality on the leave admin side.