PFMLPB-2746: Adding Leave Allotment Change Notices and Notifications to the E2E test suite (#6877)
Added a new cypress test `api_caring_continuous_historicalLeaveAllotmentChange.ts` which needs to create a historical absence case and approved case. Then will add Leave Allotment Change Notice this will be in Draft and moved to Completed. Then trigger for it has 5hr pending trigger. We are checking the notifications for both Claimant and Leave Admin to make sure they get an email. Then checking the Claimant and Leave Admin Portal for the notices to show and be able to downloaded.