INFRA-921 Resolve Lambda Security Findings (#6869)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-921

🛠 Changes
added a AWS:SourceAccount condition to lambdas
ℹ️ Context for reviewers
Security Hub is showing many different critical warnings. Fix all of them to consider this ticket complete. If there is an issue documented, but it’s the expected result, then document it in the comments and change the ‘Workflow status’ in Security Hub. Most (if not all of the issues) will either be easy fixes or have a one-fix-for-all approach, so they are being combined in one user story.

Lambdas were returning the following warning: [Lambda.1] Lambda function policies should prohibit public access. This warning is triggered because a resource-based policy statement for a lambda grants public access.