INFRA-886 Investigate release tool versioning bug (#6844)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-886

🛠 Changes
updated portal logic for versioning
ℹ️ Context for reviewers
On Dec 28, portal/v50.0 was due to be released. In the original code, there was a workaround to account for the portal's two digit versioning system. This involved adding a 0. to the front of the version number so that the library could increment/perform other actions, then splitting off the 0. before outputting the final object for branch naming and tagging. However, because the new tag had two places that contained a 0. the string manipulation only returned what would have been the minor number. (aka this weird tag portal/v0)