PORTAL-265: Add generic ApiResourceCollection model (#6946)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-265

## 🛠 Changes

- Adds new `ApiResourceCollection` model, to replace `BaseCollection` and its subclasses.
- Removes `TempFileCollection` in order to demonstrate usage of the new collection model.
- Updates `useCollectionState` to support either a `BaseCollection` or the new `ApiResourceCollection`

## ℹ️ Context for reviewers

Loren put together [a brief spec](https://lwd.atlassian.net/l/c/H1CSmLt8) for this work back in August 2021, and this PR begins to implement it. TLDR: The idea is to remove all of the `*Collection` model classes, in favor of the `ApiResourceCollection` generic class that this PR adds. It would simplify and improve our TypeScript types, and I kinda like the idea of not having to create a new class anytime we want to manage the state for a group of API resources. Some [additional context is in the ticket description](https://lwd.atlassian.net/browse/PORTAL-265).

This is a variation of the original spec, which called for also adding an `BaseApiResource`. This variance is documented in "Propsal 2" in the linked spec.