PFMLPB-3068: Adjust notification queries to search for correct NTN (#7005)
https://lwd.atlassian.net/browse/PFMLPB-2471
Subtask: https://lwd.atlassian.net/browse/PFMLPB-3068

We currently have a number of specs which query for emails which have a common subject (specifically, Application update notice from the Department of Family and Medical Leave) and look for a claim ID within the email. Because a number of legal notice notifications (both claimant and LA) use this exact subject and contain the NTN in the email body, there's the potential that we'll either erroneously mix up emails containing different notices (e.g. Leave Allotment Change and Change Request Approved notifications) or mix up claimant and LA emails. Both scenarios can introduce false positives and thus test instability.

This solution extends the messageWildcard option passed to getEmails to accept a RegExp pattern and modifier flags. That way, we can query for only the email which contains the target NTN and the copy specific to the target audience (claimant or LA) and target notice.