END: Roll back changes to E2E concurrency and addition of infra-test environment (#7022)
This is a rollback of a few select pieces of INFRA-817 (PR).
Reduces the concurrency E2E tests run with from 10 -> 7. This is intended to reduce the overload issues we've seen in Fineos. We believe this change was made unintentionally when additional environments were added, and that it was intended to allow the testing of additional environments. Instead, it's increasing the concurrency per environment.

Removes the infra-test environment from E2E runs. As far as I know, this environment does not have its own Fineos instance, so testing it would be problematic at best, and most likely impossible.