PORTAL-265: Remove AppErrorInfoCollection (#6978)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-265

## 🛠 Changes

- Replaces `AppErrorInfoCollection` with basic array and `useState`

## ℹ️ Context for reviewers

See #6946. This PR is slightly different than the others, in that it isn't replacing `AppErrorInfoCollection` with `ApiResourceCollection`. This class felt a little overkill, given the `items` getter was primarily what was used, which can be represented as a plain array. This is safer now that TypeScript is in the picture.