PFMLPB-2943 Implement lint-staged for e2e directory (#7035)
https://lwd.atlassian.net/browse/PFMLPB-2943

Implements lint-staged for the E2E directory. This will add commit-time linting and formatting of files that are changed. It should help us avoid introducing linting or formatting issues in the future.

Updates lint-staged in the root to 12.3.2. This was necessary to allow for our .lintstagedrc.json to live in the e2e directory.
Add a .prettierignore file and unify it with .eslintignore.
Split npm run lint into two things: lint (ESLint and TSC), and format (prettier). It's now possible to lint files without updating their formatting.
Add in a commented-out npm format-check to the e2e-ci workflow. When we've addressed all of the linting issues in our codebase, we can uncomment this to prevent them from being reintroduced.
Note: This doesn't actually fix any of the linting errors in our codebase aside from the formatting of jest.unit.json. That's a separate ticket that will follow.