PFMLPB-2943: Remove E2E Lint-staged RC to roll back broken linting change (#7070)
https://lwd.atlassian.net/browse/PFMLPB-2943

Rolls back the addition of an E2E lint-staged configuration. This turned out to be not work as expected, as lint-staged uses the current working directory when it runs subdirectory tasks. A fix for this is coming upstream (okonet/lint-staged#1091), but in the meantime, we want to disable lint-staged checking in the E2E directory.