API-2340: Automatically raise Unauthorized error if app.current_user() is None (#7082)
## Ticket
https://lwd.atlassian.net/browse/API-2340

## Changes
- Changed app.current_user() to return the current User or throw an exception. Originally returned an optional and allowed caller to handle the None state, but everywhere was only returning unauthorized.
- Also changed POST /employers/verifications to return an Unauthorized error instead of manually constructing a BadRequest (probably legacy copy/paste)
- Update usages of app.current_user() to not manually nullcheck and throw

## Demo
Added one unit test that explicitly verifies if calling /api/v1/users/current without an auth header, that it receives a HTTP 401 Unauthorized. It's very unlikely that app.current_user() == None in any other circumstance so we'd have to go out of our way to create a nearly impossible app state to test this more directly.