PFMLPB-3118: Refractor the OrgUnits cypress test to use in other environments (#7053)
https://lwd.atlassian.net/browse/PFMLPB-2512
Subtask: https://lwd.atlassian.net/browse/PFMLPB-3118

Instead of using the originally org unit dataset, we use the dataset from FEIN 04-6002284 (Commonwealth of MA). That has been manually setup in certain environments with the org units. Refactored the cypress test to make adjustments and moved this to the stable folder for DT3 and Perf only at this point. We won't be able to run this in DT2, IDT, and UAT at the moment.