Chore: Remove and ignore changes to pfml_api/bin/Debug directory (#7002)
## 🛠 Changes

- Adds `pfml_api/bin/Debug` directory to `.gitignore` so that its contents aren't checked into the repo

## ℹ️ Context

Some of these files were checked in as part of #6885. Some of them include file paths that are unique to the engineer's machine, like:

```
/home/joelpelaez/src/pfml/pdf_api/bin/Debug/net6.0/appsettings.Test.json
```

Given that ^ and the fact that these are in a "Debug" directory, it seems safe to ignore these so they're not committed into the repo. The impact of committing them into the repo is that it increases the repo size and every engineer then gets these changes when they pull from `main`.  This change _does not_ result in the files being removed from an engineer's local machine, just from GitHub.