PFMLPB-2181: Do not instantiate azure if the client_secret is empty (#7152)
Since the client_secret was added to the terraform container definition, the container will throw an error if it doesn't exist in parameter store. With the client_secret set, all of the azure related environment variables are set, Azure gets instantiated, and it tries to make a request to https:////discovery/v2.0/keys?appid= causing an uncaught exception.