PFMLPB-2420: Use workflow_call when invoking E2E tests (#7228)
Implements our E2E tests as a "Reusable Workflow." This is part of a broader effort to transition us to using workflow_call instead of workflow_dispatch. workflow_call is a new feature introduced by Github, where "reusable workflows" can be directly integrated into other workflows, avoiding the need for a separate trigger (which we were doing with the REST API), and allowing the results to show in "in context" for the triggering workflow.