INFRA-1127: Updating files (#7263)
* Updating README files

This is simply a change to test how GitHub workflows are executed when changes are made to multiple projects. The extra lines in the existing README files will be reverted.

* Fixing formatting issues

The whitespace broke the format verification. Replacing whitespace with updated headings.

* Reverting unneeded README changes

Changes made in the previous commit are being reverted.

Co-authored-by: Matt McCarty <mattmccarty@users.noreply.github.com>