PFMLPB-3315: Create an integration test to test the data extracts (#7268)
https://lwd.atlassian.net/browse/PFMLPB-3315

Add getFineosExtracts() to InfraClient. This new method accepts a date and uses it to find the copied extract records from that particular date.

getFineosExtracts() is used in e2e/test/integration/fineos_extracts.ts to confirm that extracts ran successfully by looking for a file that has a suffix of -OK (this file is only generated if extracts ran successfully)

Context:
Currently, our tests will assume that extracts are successful by only locating the OK file and confirming that it isn't the only file in the extract folder. This is good baseline coverage, but a list of using a list of every expected file per environment will deepen our coverage as we would be able to catch any file that is missing from a particular extract.