PORTAL-1781: Prefix generic "incorrect" messages with a unique code (#7408)
## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-1781

## ℹ️ Context for reviewers

We intentionally set a generic error response for a group of validation rules during application import, due to security concerns. A number of [the errors](https://lwd.atlassian.net/wiki/spaces/CP/pages/2219344016/Channel+Switching+Error+Messages) return a response with its error type set to incorrect, and the same message. 

To assist in debugging things though, it would be helpful to have some unique aspect of this response so engineers can understand what is incorrect about the request.