PFMLPB-3393: NR Dash - Create new environment overview (#7371)
This is the start of a full refactor for the new relic E2E Suite Dashboard. This leverages 2 new tables to allow better query structures and easily query for statistics as well as get detailed run information. The strategy is to move to new components while keeping the current components in tact. This allows us to code as needed without disrupting the current application being used as well as beta test the new views.