INFRA-1115 Move Portal Build Ci actions to Teams  (#7516)
🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1110
Subtask: https://lwd.atlassian.net/browse/INFRA-1115

🛠 Changes
removed Slack refs such as notification channel, and refs to Slack API key
refactor jobs to post results to teams instead
ℹ️ Context for reviewers
This PR is part of an effort to migrate alerts and other notifications from Slack to Teams (a tool approved for the Commonwealth). These notifications will be found in the mass-pfml-deploys-shared channel in Teams. Other similar migrated alerts will be found in Teams channels with the same name as their Slack counterparts. It is recommended that you pin this Teams thread