INFRA-1118: redirect portal-links-scanner action from slack to teams (#7501)
* replace post-failures-to-slack with post-to-teams

* fix title

* Update portal-links-scanner.yml

* INFRA-1140: Put the static site storybook behind a CloudFront domain (#7423)

* Created storybook module

* Removed duplicate local variables

* Removed response header policy

* Update storybook workflow output storybook cf id

* Added storybook to infra-test

* Updated terraform output format

* Enable terraform wrapper

* Updated checkout

* Added cloudfront logging, changed workflow from infra-test to test

* Removed cloudfront logging

* Removed obsolete code

* add secure actions to one cypress test ran in the morning (#7508)

Removing the secure actions it blocks from the cypress tests and moving them into one cypress test that will be run in the morning. This will be run in all environments during the morning runs directory. These tests are all FINEOS testings for secure action are imported correctly after a deployment.

* INFRA 1140: put the static site storybook behind a cloud front domain (#7511)

* Created storybook module

* Removed duplicate local variables

* Removed response header policy

* Update storybook workflow output storybook cf id

* Added storybook to infra-test

* Updated terraform output format

* Enable terraform wrapper

* Updated checkout

* Added cloudfront logging, changed workflow from infra-test to test

* Removed cloudfront logging

* Removed obsolete code

* Added principles to IAM statement

* Updated principle for IAM policy

* FMT

* API-2319: Screen for 26 weeks of total leave  (#7196)

https://lwd.atlassian.net/browse/API-2319

Implements payment post processor to flag payments with a leave duration exceeding 26 weeks.

* API-2436: Remove Mmmars payments backfill script (#7473)

https://lwd.atlassian.net/browse/2436

Removed the unused mmars payments backfill script, and deprecated the payment_id column from the MmarsPaymentData table.

* INFRA-1199 fix mismatched names and descriptions in sns alarms (#7507)

* update names

* update description

* PORTAL-1768: Add feature flag for syncing cognito mfa prefs (#7510)

## 🎫 Ticket

https://lwd.atlassian.net/browse/PORTAL-1768

## 🛠 Changes

* Adds a `claimantSyncCognitoPreferences` feature flag to the Portal, set to false
* When the above feature flag is set then requests to PATCH /users/:id will include a new `X-FF-Sync-Cognito-Preferences` header set to true
* Adds the `X-FF-Sync-Cognito-Preferences` header to the openapi spec for the PATCH /users/:id endpoint

## ℹ️ Context for reviewers

Future PRs will begin using these feature flags. Using these feature flags will allow us to synchronize the rollout of the logic changes that are part of this ticket's implementation across the Portal and API

## 👀 Demo

<img width="1304" alt="Screen Shot 2022-02-28 at 4 47 00 PM" src="https://user-images.githubusercontent.com/9039672/156064442-c692a1d1-0db8-423e-aa88-6b62dce3dd49.png">

## 📜 Link to relevant documentation

N/A

* API-2474: Handle `nillable` fields in FINEOS Web Service Composer responses (#7496)

Decode XML fields that have the `@xsi:nil` attribute as `None`.

This is possibly applicable to XML responses from the "integration services"
endpoints (like the `CreateOrUpdateLeaveAdmin` XSDs), but for now only using it
for the things that actually have a `nillable` attribute declared in the XSDs,
which is currently only the "wscomposer" ones.

Ticket: https://lwd.atlassian.net/browse/API-2474

* API-2400: Update payment SQL reports with payment transaction type (#7349)

* INFRA-1115 Move Portal Build Ci actions to Teams  (#7516)

🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1110
Subtask: https://lwd.atlassian.net/browse/INFRA-1115

🛠 Changes
removed Slack refs such as notification channel, and refs to Slack API key
refactor jobs to post results to teams instead
ℹ️ Context for reviewers
This PR is part of an effort to migrate alerts and other notifications from Slack to Teams (a tool approved for the Commonwealth). These notifications will be found in the mass-pfml-deploys-shared channel in Teams. Other similar migrated alerts will be found in Teams channels with the same name as their Slack counterparts. It is recommended that you pin this Teams thread

* INFRA-1116 Move Portal-CI Messages from Slack to Teams(#7495)

🎫 Ticket
https://lwd.atlassian.net/browse/INFRA-1110
Subtask: https://lwd.atlassian.net/browse/INFRA-1116

🛠 Changes
removed Slack refs such as notification channel, and refs to Slack API key
refactor jobs to post results to teams instead
ℹ️ Context for reviewers
This PR is part of an effort to migrate alerts and other notifications from Slack to Teams (a tool approved for the Commonwealth). These notifications will be found in the mass-pfml-deploys-shared channel in Teams. Other similar migrated alerts will be found in Teams channels with the same name as their Slack counterparts. It is recommended that you pin this Teams thread

* PFMLPB-3580: fix TestResultInstance tryNumber order (#7514)

* PORTAL-1813 Rename useAppErrorsLogic to useErrorsLogic (#7513)

* Rename appError* to error*

* remove image

* remove failing test

Co-authored-by: Jamie Taschetti <91554519+jtaschetti@users.noreply.github.com>
Co-authored-by: Mark Baker <95445362+markbcode@users.noreply.github.com>
Co-authored-by: Melissa Rossi <mrossi113@gmail.com>
Co-authored-by: Josh Osborne <josh@qcmmedia.com>
Co-authored-by: Sheldon Bachstein <bachsteinsk@gmail.com>
Co-authored-by: Tanner Doshier <git@doshitan.com>
Co-authored-by: Michael Chouinard <46358556+chouinar@users.noreply.github.com>
Co-authored-by: Alsia Plybeah <alsia@navapbc.com>
Co-authored-by: Zanson Watt <91619176+zansonLCM@users.noreply.github.com>
Co-authored-by: Libbey White <92495948+libnava@users.noreply.github.com>